# Recursive implementation of `searchsortedfirst` and `searchsortedlast`