--- a/src/lazyconcat.jl
+++ b/src/lazyconcat.jl
@@ -1012,7 +1012,10 @@ _searchsortedlast(a::Number, x) = 0 + (x ≥ a)
 searchsortedfirst(f::Vcat{<:Any,1}, x) =
     searchsortedfirst_recursive(0, x, arguments(f)...)
 
-@inline searchsortedfirst_recursive(n, x) = n + 1  # no more arguments
+searchsortedlast(f::Vcat{<:Any,1}, x) =
+    searchsortedlast_recursive(length(f), x, reverse(arguments(f))...)
+
+@inline searchsortedfirst_recursive(n, x) = n + 1
 
 @inline function searchsortedfirst_recursive(n, x, a, args...)
     m = length(a)
@@ -1021,13 +1024,13 @@ searchsortedfirst(f::Vcat{<:Any,1}, x) =
     return searchsortedfirst_recursive(n + m, x, args...)
 end
 
-function searchsortedlast(f::Vcat{<:Any,1}, x)
-    args = arguments(f)
-    for k in length(args):-1:2
-        r = _searchsortedlast(args[k], x)
-        r > 0 && return mapreduce(length,+, args[1:k-1]) + r
-    end
-    return _searchsortedlast(args[1], x)
+@inline searchsortedlast_recursive(n, x) = n
+
+@inline function searchsortedlast_recursive(n, x, a, args...)
+    n -= length(a)
+    r = _searchsortedlast(a, x)
+    r > 0 && return n + r
+    return searchsortedlast_recursive(n, x, args...)
 end
 
 searchsorted(f::Vcat{<:Any,1}, x) = searchsortedfirst(f, x):searchsortedlast(f,x)