@dlfivefifty 2021-01-16
Can you add a test using `@inferred`? And get the code coverage tests passing