@jipolanco 2021-01-17
I added the inference tests.

There's not much that I can do for the coverage tests, since the coverage regression concerns a single line in a file that I didn't touch (somewhere in `lazyoperations.jl`, in the `shuffle_algorithm` function), which seems to be a bit random.